revert to calling pause() in resume() but be explicit about resuming or pausing
In particular if the statue is PausedState it will now explicitly try to resume/play and if not it will try to explicitly pause. This approach still leaves thing sto be desired though since the player's media will be nil unless playInternal() gets called, so one could call pause but without actually pausing or resuming anything, while pause on a not yet played media might very well be intent to play && pause. This is to be taken into account and ultimately loading would imply buffering so that either resume or pause would have a result.