Introducing new debug magic featured by Amarok.
The key difference between this version and the one in Amarok is that this one is QDebug only.
This is currently only added locally as adding this to Phonon requires careful thinking to make sure we do not run into clashes with the old debug stuff there. Also naming needs some pimping etc.
Eventually this is supposed to become a bit of a standalone library for use across various projects, so it will re-grow KDebug support as compile time feature I suppose.