Do not discard adjust changes before video becomes available.
Due to missing event in libvlc for added video we do nasty vout
detection in the MO leading to a state where the VW has no vout to set
adjusts on. In this state the widget now adds the requested set
operation to a hash of pending sets. The hash uses the function name and
the value to call invokeMethod on them once the MO has a video.