Make Video appear quicker (aka almost instantly)!
Let us be a bit daring here. Rather than looking for video on the 15th
timechanged event, let's try 15 times and then no more. Usually I get
video on the first or second timechanged event, so there is no point in
having it delayed for as long as 15 event callbacks.