"fix" subtitle control in dragon player
It actually is more of a patch for the problem as the libphonon design is heavily flawed in that it exposes MediaController specific descriptors to the world also via the backend. Now as the Backend is a singleton but the MC is not, there is a bit of a scope problem... The present solution thus introduces a global overlord master subtitle manager where each MC needs to store its subtitles. Additionally libvlc seems broken as it reports completely bogus SPU ids \o/

The code at hand needs refactoring && bugfixing && cleanup && documentation && *whine*

BUG: 267424