Make HSB value conversion saner
Instead of messing around with percent values, make use of the fact that we either
normalize -1..1 to 0..2 or 0..1, hence a simple multiplication with either the
upper boundary or half the boudary is sufficient to scale the normalized values
to the target range.

Additionally stop setting the upper boundary by default to 2.0, it makes the calling rather confusing.