realign relationship between MO and VW
Now the MO does not need to know about the VW.
In PlayInternal we call out to all attached sinks, that they
should add stuff to the Media we have created, this is a perfect
opportunity to set the WinId for overlay painting (from inside
the VW class). That way the MediaObject does not need to know
whether there is a VideoWidget or not.