Reorganize DeviceManager lists by unification
Previously, the device lists were separate depending on the device type. But that induced code duplication and a couple
of cryptic methods. Now, the code looks cleaner.

Moved the object indexes / properties code regarding devices from the Backend to an unified place in the DeviceManager.

Additionaly, removed two unused methods.