Stop applying the volume on play/resume
When PASupport is active it intercepts all AO calls and applies them to PA
directly, so changing the volume regardless of PA will lead to
unpredictable behavior as the volume value the backend AO knows about might
not be the actual volume value in the frontend/PA, so it essentially sets
a wrong volume on play/resume.

Instead it now sets the volume when a path is created between the AO and
the MO if and only if PASupport is not active. If PASupport is active it
does not set the volume at all.

Verification pending.

BUG: 296589