Handle sound systems when libVLC gives no devices for them
libVLC offers an empty device list for sound systems such as OSS. Create
a default device for the sound system in this case.

Additionaly, if multiple sound systems are found, return devices for all
of them, not just for the first one.

Thanks to Franz Fellner for reporting the issue.

Signed-off-by: Harald Sitter <sitter@kde.org>