AudioDataOutput works fine with VLC 2.1 again (activate it in Backend)
If anyone has VLC 2.0.1 or better yet VLC 2.0 git, please feel free to
remove the #if and test if it works for you. It will probably be a while
until VLC 2.1 gets out, so if we get a 2.0 release that fixes the issue
it would be preferable.

CCMAIL: kde-multimedia@kde.org