Don't reset members all over the place
resetting too early leads to many abouttofinish signals which screws with
Amarok which likes to bind half it's playback bound features to that
particularly signal.

Hence only reset once we got a new call to PlayInternal at which point
we throw away the old media anyway, so we can be sure that this is the
last point anything from the old media might leak through.

CCBUG: 292510