partial revert of 3d7ca3f087946981301c02e96e2badce279bbc7d
The totalTime cache is needed because we need to provide -1 as totalTime()
for as long as we do not get a proper update through this slot.
VLC reports -1 with no media but 0 if it does not know the duration, so
apps that assume 0 = unkown get screwed if they query too early.

For Phonon 4 we better stay safe than sorry.
Phonon 4.7's documentation now explicitly states that totalTime must only
be taken serious after Playing state was reached or its signal was emitted.

http://bugs.tomahawk-player.org/browse/TWK-1029