simplify ticking logic
upon further review of the qtphonon documentation it is not really
clear how ticking is supposed to behave precisely, however it seems
to be closest when time stamps are not duplicated, so timer ticking is
out of the question as that can potentially do that since we have no
control over VLC's time stamp resolution (and I also found no way to
actually get any value of the resolution thus actually rendering us
partially incompatible with the documentation's requirements \o/).

Instead of time ticking we simply take whatever VLC gives us and check
the delta to the last tick we emitted if it is >interval we emit a new tick
otherwise we ignore the time change.