implement pausedplay mechanic
phonon's api suggest that pausing before playing is possible (through
naming of the interfaces) however vlc does not provide any such
functionality so once again we do a nasty intercept-hack.

pausedPlay issues a regular play but then intercepts the VLC playing
event and directly switches to paused.
nothing outside MediaPlayer ever knows that VLC actually was in playing
state and a completely transparent switch from non-playing to paused
becomes possible.

please note that this does not in any way guarantee that no audio/video
samples are already sent to the output. it merely tries to reduce the
chance of this happening by calling pause in the event callback, which
is of course threaded from the pipeline so anything can happen between
us getting the signal and asking VLC to pause.

verification for backport pending.

BUG: 305028