protect MediaPlayer::setCdTrack against bogus m_media accesses
this *probably* caused a crash when trying to set a track on a newly
created MO. m_media up until now was not initialized to 0 so it is
possible that setCdTrack was run on an invalid Media object.
outside this init case there is no point where either m_media or the
underlying vlc media would silently get a bogus value (see MP::setMedia,
Media::Media).

BUG: 329121
FIXED-IN: 0.7.2