Prevent implicit play calls when having queued an invalid source
Some applications (knotify4) queue an intentionally bogus source as
last source to ensure that we terminate playback. This had the opposite
effect since we un-broke queuing as now we'd do a bogus call for play()
after not having changed the source. This was causing a chance
(somewhat timing dependent) to get duplicated sounds at the end of
playback.

moveToNextSource now checks whether m_nextSource actually is valid and
only if it calls play().

(in the long run it may be worthwhile to consider having setSource return
 a bool, to only call play() iff setSource actually was able to do stuff)

BUG: 330154
FIXED-IN: 0.7.2