thorw away previous libvlc loggin and use QLoggingCategory instead
previously we'd try to log to a file (which never worked) and asked for
direct printf'ing which did work but was overly out of line with the rest
of debugging. fortunately libvlc introduced logging callbacks and qt
introduced debug categorization, so we can now use that.
this does two things
a) mostly deprecate the previously used debug envrionment vars because
   one can use qlogging.ini to define the verbosity (whether we want to
   keep the env in addition to that is yet to be decided)
b) passes all of libvlc's debug through qdebug allowing it to be neatly
   integrated with all its abilities

for configuration reasons we use a different category for the subsystem
than for the backend (as before with two env vars)