do not attempt to paused-play pause an unpausable player
if a player can not be paused, libvlc internally calls stop on the player
when one attempts to pause it anyway. since stop tears down the ES it is
somewhat blocking and releases resources. if a phonon application then
causes another stop call the application can end up in a nice and dandy
deadlock as two threads are now waiting on resources they will not ever
get.
to prevent this the libphonon pause behavior is now getting bent to the
libvlc behavior (since we have no real way around it anyway short of
implementing a fake pause which would be entirely simulated by phonon-vlc).

long story short: using pause can now explicitly result in a stopped state
when using a media that can not be paused; such as certain stream types.

BUG: 337604
REVIEW: 119370