vlc 2.2 compatibility - do not use pulsesupport overlay
In vlc 2.2 the pa_stream creation time was changed such that the original
workaround for bug 321288 (todo with pulseaudio property hijacking for
the overlay) has adverse effects in that the overlay would be defunct
because the properties were no longer set before stream creation.
Given the nature of the workaround there is no suitable solution to this
other than expanding the scope of the environment overrides again which
would then mean that pvlc apps can once again not use multiple players.

Because this pile of crap, the pulse overlay is now entirely
disabled by the backend when built against vlc >= 2.2.

since vlc 2.1 already supported a bunch of things natively that the
overrides did previously, here is a list of known and accepted regressions:
- changing the volume of the pulse stream is not propagated to phonon
- changing the muteness of the pulse stream is not propagated to phonon
- the media.role is always 'video' making all phonon applications use
  the video category for device selection (on the PA side of things)

The KCM should not be affected by this as it has its own device listing
implementation outside pulsesupport. However, all applications that
directly rely on Phonon for device listing will in all likelyhood get
backend device listing (i.e. what libvlc knows) instead of pulse devices
which in turn renders device control moot on pulse systems as libvlc
can not control pulse sinks.