hide a bug whereby libvlc sends vout events but then has_vout is off
happens with 4.0-dev fairly regularly and messes up dragon as it'd get
the event and then still query hasVideo which will be false when the
event was true. not quite clear why it happens, but supposedly is some
state confusion inside the media player struct (in particular since we
only dispatch the notification to the frontend in the next eventloop cycle,
so it can't even be a problem with us being too quick. i.e. the
event callback definitely returned by the time we even become aware of
the change)