port controller away from vlc 3.0 deprecations
chapter descriptions and title descriptions are no longer linked lists but
carrays. to conveniently handle that, move away from the pre-processor
iterator to a simple container for the array which helps with recycling.

arguably we maybe should make this a range loopable container, not useful
for now as we only need the index just now though.