stop using our helper function to calculate pitch and lines
the surface painter was more inefficient than it needed to be by
manually calculating a pitch and stride but then constructing a qimage
anyway. after some research it seems that we can simply construct our
qimage and pass its picture values to VLC with the assumption that VLC
will adjust the data as it gets passed into vmem output (i.e. our data
blob). this then also allow us to get rid of the intermediate plane
bytearray and simply use the memory the qimage allocates anyway.

CCBUG: 434506