Fix building and installing translations
This broke in [1]. While ecm_create_qm_loader can take a target, it
just doesn't exist at the time of invocation. Use the opportunity
and introduce a more target-centric approach to address this.
Furthermore, after the move of ecm_install_po_files_as_qm() to
src/CMakeLists.txt (to avoid duplicate targets), the path passed to it
needs adjustment.

[1] 8b335d872994196d1dab6f12a3b45bfc2427d910